refactor(server): extract device & system route groups (Refs #3502)
Continue the server.ts route extraction (issue #3502). Moves two
cleanly-decoupled handler groups out of the inline monolith.

device routes (routes/deviceRoutes.ts), mounted at '/':
  GET  /device-config
  GET  /device/backup
  POST /device/reboot
  POST /device/purge-nodedb

system routes (routes/systemRoutes.ts), mounted at '/':
  GET  /system/status
  GET  /status
  GET  /version/check
  POST /system/restart

The Docker/version pure helpers (isRunningInDocker, compareVersions,
checkDockerImageExists) plus serverStartTime move to a shared
utils/systemInfo.ts, since the startup auto-upgrade scheduler
(checkForAutoUpgrade) still references compareVersions/checkDockerImageExists.
The restart handler's server-lifecycle gracefulShutdown is injected via
setSystemCallbacks() so the route module stays free of the HTTP-server ref.

Handler bodies moved verbatim; middleware preserved. Orphaned
deviceBackupService/backupFileService imports removed from server.ts.

Inline apiRouter handlers: 98 -> 90.

Adds routes/deviceRoutes.test.ts and rewrites routes/systemRoutes.test.ts
as supertest coverage of the new modules.
